move source to src, tests to tests, make use of composer psr-4
