+++ phpunit.xml.dist
@@ -1,8 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
-<phpunit colors="true" bootstrap="./vendor/autoload.php">
+<phpunit
+        bootstrap="vendor/autoload.php"
+        colors="true"
+>
     <testsuites>
         <testsuite>
-            <directory>./Tests</directory>
+            <directory>tests</directory>
         </testsuite>
     </testsuites>
 </phpunit>
