+++ src/DataStructureLengthSniff.php
