+++ src/Sniffs/Classes/PropertyVisibilitySniff.php
