+++ src/Sniffs/CodeAnalysis/ArrayPropertyPerClassLimitSniff.php
