+++ src/Sniffs/CodeAnalysis/OneObjectOperatorPerLineSniff.php
