+++ src/Sniffs/ControlStructures/NoElseSniff.php
