+++ src/Sniffs/Files/FunctionLengthSniff.php
