+++ src/Sniffs/Files/InterfaceLengthSniff.php
