+++ src/Sniffs/Files/TraitLengthSniff.php
