+++ src/Sniffs/NamingConventions/ClassLengthSniff.php
