+++ src/Sniffs/NamingConventions/ConstantLengthSniff.php
