+++ src/Sniffs/NamingConventions/FunctionLengthSniff.php
