+++ src/Sniffs/NamingConventions/VariableLengthSniff.php
