+++ tests/Sniffs/Classes/PropertyVisibilityUnitTest.inc
