+++ tests/Sniffs/CodeAnalysis/ArrayPropertyPerClassLimitUnitTest.inc
