+++ tests/Sniffs/CodeAnalysis/ArrayPropertyPerClassLimitUnitTest.php
@@ -1,15 +1,15 @@
 <?php
 
-namespace ObjectCalisthenics\Tests\NamingConventions;
+namespace ObjectCalisthenics\Tests\Sniffs\CodeAnalysis;
 
 use ObjectCalisthenics\Tests\AbstractSniffUnitTest;
 
 /**
- * Constant length, part of "Do not abbreviate" OC rule test.
+ * Array property per class limit, part of "Use first class collections" OC rule.
  *
  * @author Guilherme Blanco <guilhermeblanco@hotmail.com>
  */
-class ConstantLengthUnitTest extends AbstractSniffUnitTest
+class ArrayPropertyPerClassLimitUnitTest extends AbstractSniffUnitTest
 {
     /**
      * Returns the lines where errors should occur.
@@ -22,7 +22,8 @@ class ConstantLengthUnitTest extends AbstractSniffUnitTest
     public function getErrorList()
     {
         return array(
-            9 => 1,
+            3  => 1,
+            24 => 1,
         );
     }
