+++ tests/Sniffs/CodeAnalysis/InstancePropertyPerClassLimitUnitTest.inc
