+++ tests/Sniffs/CodeAnalysis/OneObjectOperatorPerLineUnitTest.inc
