+++ tests/Sniffs/Metrics/MethodPerClassLimitUnitTest.inc
