+++ tests/Sniffs/NamingConventions/ConstantLengthUnitTest.inc
