+++ tests/Sniffs/NamingConventions/VariableLengthUnitTest.inc
